// src/app/api-client/model/clinicalPatientMatrixExportFilterModel.ts
/**
 * WebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ClinicalPatientMatrixFilter } from './clinicalPatientMatrixFilter';

export interface ClinicalPatientMatrixExportFilterModel { 
    filter?: ClinicalPatientMatrixFilter;
    currentPage?: number;
    pageSize?: number;
    fileName?: string;
}